Math.NumberTheory.Moduli.Jacobi

Description

Jacobi symbol.

Documentation

data JacobiSymbol

Type for result of jacobi.

Constructors

MinusOne

Zero

One

jacobi :: (Integral a, Bits a) => a -> a -> JacobiSymbol

Jacobi symbol of two numbers. The "denominator" must be odd and positive, this condition is checked.

If both numbers have a common prime factor, the result is 0, otherwise it is ±1.

jacobi' :: (Integral a, Bits a) => a -> a -> JacobiSymbol

Jacobi symbol of two numbers without validity check of the "denominator".